requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# azure-functions
# azure-functions-durable
azure-functions==1.11.2
azure-functions-durable==1.1.6
azure-storage-blob==12.12.0
pandas==1.4.3
numpy==1.22.4
faker